✨ 스마트컨트랙트를 통한 리워드 지급 로직 추가 (#24)
* ✨ presigned url 발급 기능 추가

* ✨ 스마트컨트랙트를 통한 리워드 지급 로직 추가

* 🎨 포매팅 적용

* 🎨 타입 설정 변경

+++ app/schemas/store.py
@@ -30,7 +30,7 @@ class StoreEventSummary(BaseModel):
     id: str
     title: str
     condition: str
-    reward: int
+    reward: float  # ETH
     isActive: bool
